version: move to internal/version
