+++ internal/cmd/ansible-operator/run/cmd.go
@@ -41,7 +41,7 @@ import (
 	"github.com/operator-framework/operator-sdk/internal/ansible/runner"
 	"github.com/operator-framework/operator-sdk/internal/ansible/watches"
 	"github.com/operator-framework/operator-sdk/internal/util/k8sutil"
-	sdkVersion "github.com/operator-framework/operator-sdk/version"
+	sdkVersion "github.com/operator-framework/operator-sdk/internal/version"
 )
 
 var (
