+++ internal/plugins/ansible/v1/scaffolds/internal/templates/dockerfile.go
@@ -20,7 +20,7 @@ import (
 	"sigs.k8s.io/kubebuilder/pkg/model/file"
 
 	"github.com/operator-framework/operator-sdk/internal/plugins/ansible/v1/constants"
-	"github.com/operator-framework/operator-sdk/version"
+	"github.com/operator-framework/operator-sdk/internal/version"
 )
 
 var _ file.Template = &Dockerfile{}
