+++ release.sh
@@ -34,7 +34,7 @@ if ! go version | cut -d" " -f3 | grep -q "$GO_VER"; then
 fi
 
 # Detect whether versions in code were updated.
-VER_FILE="version/version.go"
+VER_FILE="internal/version/version.go"
 CURR_VER="$(sed -nr 's|\s+Version\s+= "(.+)"|\1|p' "$VER_FILE" | tr -d ' \t\n')"
 if [[ "$VER" != "$CURR_VER" ]]; then
 	echo "version is not set correctly in $VER_FILE"
